backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
python-dotenv==1.0.1
pytz==2024.2
aiosqlite==0.20.0
pydantic==2.10.4
